Allow dynamic recording URLs
On creation ensure all recordings have a valid security ID token (sid).   On a new connection attempt to re-authenticate the sid  keeping saved URL's  valid.   There are cached json file for the recordings so if there is a failure recordings may not play until the 3hr cache timeout is met.  The change of a failure is low.  Unfortunately scanning tasks do not remove the cached files.

The modification date of the recording is the date of a sid change and it is also persisted so recordings are not rescanned each load.